pyproject.toml
[project]
name = "bilibili-downloader"
version = "0.1.0"
description = "Bilibili動画ダウンローダー - Web GUIアプリケーション"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "qrcode[pil]>=8.2",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
bilibili-downloader = "bilibili_downloader.app:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bilibili_downloader"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
